Refactor: 카테고리 관련 작은 오류 수정 20230925 0229

+++ Back-End/src/main/java/com/conseller/conseller/category/mainCategory/MainCategoryRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface MainCategoryRepository extends JpaRepository<MainCategory, Integer> {
-    Optional<MainCategory> findByMainCategoryIdx(long mainCategory);
+    Optional<MainCategory> findByMainCategoryIdx(int mainCategory);
 }
