+++ Back-End/src/main/java/com/conseller/conseller/category/subCategory/SubCategoryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface SubCategoryRepository extends JpaRepository<SubCategory, Integer> {
-    Optional<SubCategory> findBySubCategoryIdx(long subCategory);
+    Optional<SubCategory> findBySubCategoryIdx(int subCategory);
 }
